Start of game
This is a base that we an work off of for the game.
It contains scripts for a multiple choice format that tells you fi you are right or wrong.
Questions can be programed in and will be chosen randomly.
Currently all answer choices are in the same place. Eventually it will be good to find a way to randomize this.
The player will take damage if they get a question wrong and lose if their health goes below zero.
The two buttons on top are place holders for future code.